Extract Nop driver
