+++ src/test/java/com/github/loadtest4j/loadtest4j/driver/NopDriver.java
@@ -1,12 +1,12 @@
-package com.github.loadtest4j.loadtest4j.drivers.nop;
+package com.github.loadtest4j.loadtest4j.driver;
 
-import com.github.loadtest4j.loadtest4j.DriverRequest;
 import com.github.loadtest4j.loadtest4j.Driver;
+import com.github.loadtest4j.loadtest4j.DriverRequest;
 import com.github.loadtest4j.loadtest4j.DriverResult;
 
 import java.util.Collection;
 
-class Nop implements Driver {
+public class NopDriver implements Driver {
     @Override
     public DriverResult run(Collection<DriverRequest> requests) {
         return new DriverResult(0, 0);
