+++ src/test/resources/loadtest4j.properties
@@ -1,1 +1,1 @@
-loadtest4j.driver = com.github.loadtest4j.loadtest4j.drivers.nop.NopFactory
+loadtest4j.driver = com.github.loadtest4j.loadtest4j.driver.NopDriverFactory
